Fix add buttons
