+++ src/components/Buttons/index.js
@@ -5,5 +5,4 @@ export * from './GitHubButton'
 export * from './ZoomButtons'
 export * from './IconButtonStyled'
 export * from './SettingsButton'
-export * from './AddNodeAfter'
-export * from './AddNodeBefore'
+export * from './AddNodeButtons'
